refactor: Move common minio logic into a new file
